faust_debug/requirements.txt
# faust
faust-streaming